# openapi/chatgpt-action.yaml
openapi: 3.1.0
info:
  title: Tiiextension GPT Control API
  version: 1.3.0
  description: Control Tiiextension native runtime through a Cloudflare Tunnel.
servers:
  - url: https://YOUR-CLOUDFLARE-TUNNEL.trycloudflare.com
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ToolCall:
      type: object
      required: [tool]
      properties:
        tool:
          type: string
        args:
          type: object
          additionalProperties: true
        task_id:
          type: string
        call_id:
          type: string
        idempotency_key:
          type: string
...
paths:
...
  /v1/tools/call:
    post:
      operationId: tiiextensionToolCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCall"
      responses:
        "200":
          description: Tool result.
...